Upgrade to ember 2.18 (#8)
* Upgrading to ember 2.18.0

This commit consists of the mechanical portion of applying the
differences at
https://github.com/ember-cli/ember-addon-output/compare/v2.13.1...v2.18.0

Further testing should be done to check whether other changes would
need to be applied in order to make this addon operate perfectly.

* Fixes for common test case issues, but not ESLint

These are mechanical changes which needed to be done.  Fixes for
setting array- and object-properties seem to cause issues with the
current implementation.  We'll fix this in a separate commit.

* Only set defaults when they're not provided to the component

It may be that some properties are left unset.  We provide sensible
defaults in those cases.  It is currently desired to set these
properties in the init for complex objects, instead of defining them
in the class template.  Being in line with these hints doesn't make
the component easier to define, but staying with current best
practices makes things easier to consume afterwards.